feat(win): switch to using the native executable if available
Previously, we always used lspjs on Windows, without any way to use the
native executable on Windows. This commit changes the behavior to be
similar to other platforms where we try to use the native executable if
we are able to find it. Note that the native binary executable is not
yet packaged with the Windows extension, and so we would fallback to
trying to use lspjs if a semgrep installation isn't discovered.