release: copy exe and DLLs into Windows extension
Bundle the `semgrep-core-proprietary` Windows executable into the
extension like on other platforms. We don't build an ARM compatible
binary and only download it on win32-x64 target.

On Windows, we also need the DLLs to be able to run the executable. The
DLLs are available in the Python wheels uploaded to PyPI and we bundle
them into the Windows extension.