release: copy exe and DLLs into Windows extension (#215)
Bundle the `semgrep-core-proprietary` Windows executable into the extension like on other platforms. We don't build an ARM compatible binary and only download it on win32-x64 target.

On Windows, we also need the DLLs to be able to run the executable. The DLLs are available in the Python wheels uploaded to PyPI and we bundle them into the Windows extension.

Test plan: 

- Dry run of the publish action for this commit is available [here](https://github.com/semgrep/semgrep-vscode/actions/runs/14918203888)

PR checklist:

- [x] Purpose of the code is [evident to future readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [x] Tests included or PR comment includes a reproducible test plan
- [x] Documentation is up-to-date
- [x] A changelog entry was for any user-facing change
- [x] Change has no security implications (otherwise, ping security team)