feat(win): Switch to using the native executable on Windows (#214)
Previously, we always used lspjs on Windows, without any way to use the
native executable on Windows. This commit changes the behavior to be
similar to other platforms where we try to use the native executable if
we are able to find it. Note that the native binary executable is not
yet packaged with the Windows extension, and so we would fallback to
trying to use lspjs if a semgrep installation isn't discovered.

Also, displays a user warning about Windows support being experimental. 

Test plan:
Run the extension in debug/verbose mode and note the executable being used.

PR checklist:

- [x] Purpose of the code is [evident to future readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [x] Tests included or PR comment includes a reproducible test plan
- [x] Documentation is up-to-date
- [x] A changelog entry was for any user-facing change
- [x] Change has no security implications (otherwise, ping security team)

If you're unsure about any of this, please see:

- [Contribution guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located here](https://semgrep.dev/docs/contributing/contributing/)