fix: Use semgrep from PATH if available (#213)
Previously, we tried to use semgrep from the PATH only if the user configured an incorrect path value. This commit changes that to use semgrep on PATH even when the user has not specified a path for semgrep.

Test plan: 

Configure the extension to run semgrep in the verbose/debug mode and check that the executable used is the correct one. 

PR checklist:

- [x] Purpose of the code is [evident to future readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [x] Tests included or PR comment includes a reproducible test plan
- [x] Documentation is up-to-date
- [x] A changelog entry was for any user-facing change
- [x] Change has no security implications (otherwise, ping security team)

If you're unsure about any of this, please see:

- [Contribution guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located here](https://semgrep.dev/docs/contributing/contributing/)